requirements.txt
boto3==1.34.0
google-cloud-storage==2.10.0